requirements.txt
pyrogram>=2.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
